requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-decouple==3.8
django-cors-headers==4.3.1
Pillow==10.0.1
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
celery==5.3.6
redis==5.0.1
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
